[wicket] Rack overview: display actual sled presence (#3590)
On rack2, before:


![wicket-before](https://github.com/oxidecomputer/omicron/assets/1435635/f1a94a7b-3a6e-4be5-b609-a9a9a147b5cb)

and after:


![wicket-after](https://github.com/oxidecomputer/omicron/assets/1435635/4be75a2e-c1b8-45d6-ad2d-fc99be5877b1)

The red styling for switch 0 matches `ComponentPresence::Suspicious` in
the PR, which means either "ignition says present but we can't talk to
the SP" or "ignition says not present but we _can_ talk to the SP". For
this switch we're in the latter case: this switch revision only reports
the _other_ switch as an ignition target, not itself. (Note that we do
see the _other_ switch as present and drawn as expected.) That has been
fixed in future sidecar revisions
(https://github.com/oxidecomputer/hardware-sidecar/issues/735).

Fixes #3481.