[wicket] use Ctrl-Alt-K for rack setup (#3583)
Turns out Ctrl-Alt-S is bound to other commands in two separate
terminals. This is unfortunate, and apparently its use in Emacs wasn't
enough to stop terminals from rebinding it.

I verified that Ctrl-Alt-A, R and K are not bound to anything in
commonly-used terminals on Linux. (On Mac this is less of a problem
because terminals use the Cmd key for their shortcuts.)

Fixes #3510.