[Bootstore] Persist Fsm::State (#3598)
Persist the lrtq state via a `Ledger` and allow reloading from the
Ledger.

This change also creates a `Ledgerable` type for storing replicated
network information, but it is not yet used. Support is coming in a
follow up.

We also modify `Ledger` to use JSON instead of toml for better
generality and commonality with our other tools.