Allocate Crucible regions randomly across zpools
Currently nexus allocates crucible regions on the least used datasets.
This leads to repeating failures (see #3416 ). This change introduces
the concept of region allocation strategies at the database layer.
Currently only one strategy is used: the random strategy. However this
could be expanded.

The random strategy picks 3 distinct datasets from zpools with enough
space to hold a copy of the region being allocated. datasets are
shuffled using the md5 of the dataset UUID appended to some value. This
value can be specified to get a deterministic allocation, mainly for
test purposes, but in prod it simply uses the current time in
nanoseconds. Because the md5 function has a uniformly random output
distribution, sorting on this provides a random shuffling of the
datasets, while allowing more control than simply using `RANDOM()`.

Due to SQL limitations, to the best of my knowledge, we cannot make the
selection both random and distinct on the zpool or sled IDs, therefore
an allocation may select multiple datasets in the same sled or zpool.
But, we can detect that this has occurred, and fail the query. I have
included some code which will retry a few times if two datasets on the
same zpool are selected, as a future shuffling is likely to result in a
good selection. Note that currently in production, we don't ever have two
Crucible datasets on the same zpool, but it was raised as a future
possibility.

I have not done anything to ensure we are spread across 3 sleds. We
could attempt to implement that using the same retry-approach, but it
feels a bit hacky to me as-is, and feels moreso if we were to rely on it
for that. Regardless of how we solve the problem of distributing across
3 distinct sleds, we should plumb the allocation strategy through more
parts of Nexus when moving to a 3-sled policy so that we can relax it
to a 1-sled requirement for development/testing.

Testing whether the allocation distribution is truly uniform is
difficult to do in a reproducible manner in CI. I made some attempts at
doing some statistical analysis, but to get a fully deterministic region
allocation we would need to allocate all the dataset Uuids
deterministically, which would require pulling in a direct dependency on
the chacha crate, and then hooking that up. Doing analysis on anything
other than perfectly deterministic data will eventually result in false
failures given enough CI runs. That's just the nature of measuring
whether the data is random. Additionally, a simple chi analysis isn't
quite appopriate here since the 3 dataset selections for a single region
impact each other due to the requirement for distinct zpools.

ANYWAYS. I ran 3 sets of 3000 region allocations, each resulting in
9000 dataset selections across 27 datasets. I got these distributions,
counting how many times each dataset was selected.

```
[351, 318, 341, 366, 337, 322, 329, 328, 327, 373, 335, 322, 330, 335, 333, 324, 349, 338, 346, 314, 337, 327, 328, 330, 322, 319, 319]
[329, 350, 329, 329, 334, 299, 355, 319, 339, 335, 308, 310, 364, 330, 366, 341, 334, 316, 331, 329, 298, 337, 339, 344, 368, 322, 345]
[352, 314, 316, 332, 355, 332, 320, 332, 337, 329, 312, 339, 366, 339, 333, 352, 329, 343, 327, 297, 329, 340, 373, 320, 304, 334, 344]
```

This seems convincingly uniform to me.