sled-agent: Setup dump dev on M.2, savecore to U.2 (#3586)
Each time the sled-agent upserts a new disk, it enumerates every disk it
knows about. If an M.2 is found, it runs dumpadm(8) to set it as the
dump device. If a U.2 is *also* found, it invokes savecore(8) to
save the previous dump on the M.2, if any, to an encrypted zvol
on the U.2, and mark the dump slice as empty.

This is a bit of a kludge due to savecore(8) not yet exposing a clear
way to save-and-clear a dump *partition* other than the one configured
system-wide by dumpadm. While redundant, this is for our purposes
idempotent - as long as *any* M.2 is configured, dumps have a place to
go, and as long as any U.2 is present, we will attempt to save any
yet-unsaved kernel core dumps on all known M.2s to it.

(see RFD 118 for details of partition layout) (#2450, #2478)