Sled Agent destroys crypt/zone filesystems if it hasn't seen them before (#3642)
This builds on https://github.com/oxidecomputer/omicron/pull/3638 , with
a slightly more fine-tuned mechanism.

- `crypt/zone` is created on new U.2s, if it doesn't already exist
- If it **does** exist, then Sled Agent needs to make a call: Should I
delete the dataset, possibly containing zones from last boot, or should
I leave it alone (for idempotency)?
- In #3638 , we used "was the filesystem previously mounted" as a proxy
for this information. However, in cases where the sled agent reboots,
this information gets out-of-date -- the filesystem could be mounted,
but sled agent should remove the corresponding dataset.

This PR introduces a filesystem property, simply called 'oxide:agent',
which is a randomly generated string created statically during the
lifetime of the Sled Agent. If the sled agent reboots, this value will
be regenerated, which instructs the sled agent to wipe `crypt/zone` and
re-create it.