Put process core dumps onto the U.2 debug zvol
(Part of #2478)

This configures coreadm to put all core dumps onto the M.2 'crash'
dataset, and creates a thread that rotates them all onto a U.2 'debug'
dataset every 5 minutes.

This also refactors the dumpadm/savecore code to be less redundant and
more flexible, and adds an amount of arbitrary logic for e.g. deciding
the U.2 onto which to save cores.