ci/tuf-repo: fix dvt-dock fetch (#3706)
Adding `--depth 1` here broke clones when tools/dvt_dock_commit is not
exactly equal to whatever the HEAD of the dvt-dock repo is, because I
didn't ask GitHub for that commit, and also `git clone -b` doesn't work
that way anyway. This awful four-liner is equivalent to what I want
though, and avoids fetching the full history of the repository.