LRTQ based disk encryption (#3652)
This PR adds bootstore/lrtq support by launching a tokio task running a
`bootstore::Node` in the bootstrap agent and returning a
`bootstore::NodeHandle` for interacting with the bootstore. If there are
at least 3 sleds in the initial group of sleds sent from wicket to RSS
then trust quorum is initialized via `bootstore::NodeHandle::init_rack`
and disk encryption keys for the U.2 drives will be derived from the
rack secret. If there are fewer than 3 sleds, then the current hard
coded secret will be used to derive disk encryption keys.

We don't know whether we are going to be using trust quorum or not
before RSS is run, which means that we don't know how to generate disk
encryption keys until rack initialization. Therefore we must choose our
secret retriever dynamically. And if we end up using trust quorum, we
must wait for initialization so we have the key shares needed to
recompute the rack secret and derive the disk encryption keys. Because
zones are on U.2s we cannot bring up any zones until the rack is
initialized. However, we can't initialize the rack without the switch
zone, which is where wicket, wicketd, MGS, and RSS run. Therefore this
PR also puts the switch zone back on the ramdisk so we can initialize
the rack and bring up the other zones.