Put process core dumps onto the U.2 debug zvol (#3677)
(Part of #2478, continued in #3713)

This configures coreadm to put all core dumps onto the M.2 'crash'
dataset, and creates a thread that moves them all onto a U.2 'debug'
dataset every 5 minutes.

This also refactors the dumpadm/savecore code to be less redundant and
more flexible, and adds an amount of arbitrary logic for e.g. picking
the U.2 onto which to save cores.