Abandon contract after running command in zone
When sled agent runs a command in the context of a non-global zone, it
creates a new process contract for the child that will enter the zone.
This contract is still owned by sled-agent and, once the child
completes and exits, hangs around.

This change cleans up by abandoning the contract once the child exits
which causes it to be destroyed since the contract is not flagged in
such a way that it can become an orphan.

Fixes #3753