[RSS] Change `uplink_ip` to `uplink_cidr` in rack network config (#3755)
This is necessary (but not sufficient) to address #3588 - tfport needs
the netmask/prefix for each uplink, which means we need to collect it at
RSS time.

Most of this PR is various openapi plumbing changes. There are a couple
spots that warrant a closer look (although they are small); I'll call
them out in comments below.