Update dendrite and maghemite auto-prs (#3740)
Updates the creation / update mechanism that is used for creating
dendrite and maghemite update PRs. Previously only GitHub `push` events
could be used as a notification source. This caused an issue as the
artifacts necessary for updating were still in the process of being
built by a buildomat job. To work around this a long (1 hour) debounce
delay was used so that by the time PR creation logic ran, the artifacts
should have been generated.

Now that `check_run` events are available, they can be used to more
accurately specify when PR creation should run. This update changes the
dendrite and maghemite listeners to listen for the completion of the
`image` buildomat job in their respective repos.

Resolves #3259