Fix for stranded connection initiators (#3771)
There was a bug where the peer checked to see if it was an initiator via
an invalid check where it compared sorted `Baseboard`s, rather than
comparing addresses. This resulted in all the `StaleXXX messages` and
made it so that the peer fsm would not actually discover that it was
connected on the initiating side because of an early return. This would
result in retry messages not getting sent from the FSM. This is the most
likely cause of the RSS failure on dogfood as error messages occur
around rack init time.

While looking at the code I also noticed two other bugs, that I don't
believe we've necessarily hit. These are loss of tracking connections
when an *actual* stale message is received.