[nexus] Implement on-boot DB Schema Migration (#3760)
This PR does the following:

- Adds a `schema/crdb` directory, containing `dbinit.sql`, `dbwipe.sql`,
and a collection of versions.
- When new DB versions are added, we can update `dbinit.sql` (as we have
been doing) but also add a `up.sql` file to a directory named with the
new DB version.
- On boot, nexus acquires a lease for the ability to perform schema
changes, and it executes them. Nexus refuses to boot until the on-disk
schema version matches the version embedded in the Nexus binary.

Furthermore, we provide an example "schema migration" from version 1.0.0
to version 1.0.1.

Some implementation details:
- This PR adds several tests to confirm this behavior works as expected,
across all current and future updates
- This PR makes `db_metadata` more strongly-typed, so make it easier to
access the underlying data
- This PR modified `dbinit.sql` to be idempotent, to be in-line with the
recommended pattern for all subsequent schema migrations too
- We embed the schema changes into the Nexus binary, so they can be
parsed and selectively applied. Along these lines, it might make sense
to *stop* embedding a schema into the CRDB zone, and leave the
initialization entirely under the control of Nexus.