Queue disks for upsert when a transient error occurs (#3800)
Check every 5 seconds if there are any queued disks and we have a ready
key manager, meaning that this replica has a key share or it is using
the `HardcodedKeyRetriever`. If there are disks that need upserting then
try to do that. This allows not just for cold boot but for hot insert of
new disks when the trust quorum may not be available.

Current the only transient error is a `KeyManager::Error` which
indicates that the trust quorum is unavailable becaue not enough sleds
are online.

Fixes https://github.com/oxidecomputer/omicron/issues/3789