Exclude packages that are built into the ramdisk from the control plane image (#3803)
Currently on `main`, the `control-plane.tar.gz` in the TUF repo includes
a few tarballs it doesn't need (because they're already bundled into the
OS ramdisk). This PR prunes them out explicitly when building the TUF
manifest. This doesn't really feel like a sustainable way to do this -
maybe it should be an allowlist instead of a denylist? Or maybe there's
already a better way to do this but I don't understand `package.sh` well
enough to see it.

---------

Co-authored-by: Andy Fiddaman <andy@oxidecomputer.com>