Do not attempt to delete dataset used for swap device (#3830)
Fixes #3651 by skipping the swap dataset where we fetch datasets that
can be deleted. In the limit, we may want to log a warning here (or even
try to delete the device), but I'd like to unblock folks who are hitting
this issue.

I also updated the gimlet-standalone config to configure a reservoir and
a swap device (to round out fixing #3743).

I tested this on dunkin using a non-gimlet configuration. I ensured
there were no swap devices, installed omicron and saw that it created
one, then did an uninstall to ensure it completed successfully.