Rework locking to avoid holding instance-map lock for a long time
Rather than making the instance-map lock async, and holding it for the
duration of the zone bundle operation, we keep it sync. Instead, clone
the instance itself, which is pretty cheap since it's in an `Arc`. This
keeps the code simpler and should improve overall throughput.

Also makes a small change to the zone bundle filename for zone-specific
commands, to ensure they don't collide.