[wicket] Disable updating the sled/switch where wicketd is running (#3817)
We already have server-side rejection of updating the sled where wicketd
is running, but we have neither:

1. server-side rejection of updating the switch connected to the
scrimlet where wicketd is running
2. client-side rejection of updating the sled / switch where wicketd is
running

Rejecting updates of the switch we're connected via is correct today,
because resetting the sidecar currently requires resetting the scrimlet
it's attached to. That might not always be the case, so I'm a little
hesitant to do the work to implement 1. But after #3810 it's really easy
to do 2. This PR changes the update pane to not even give the option to
update the sled or switch where we're currently connected:


![wicket-no-update](https://github.com/oxidecomputer/omicron/assets/1435635/680bd900-513c-4580-afea-cc464f9918a7)

Fixes #3505.