add rack-topology package feature
This adds the rack-topology package feature with possible values of
single-sled or multi-sled. single-sled is intended for dev/CI
deployments, while multi-sled is intended for dogfood/prod. The value
of this determines which nexus config-partial.toml is packaged.

Right now the only difference single/multi is the crucible region
allocation strategy.