ci/tuf-repo: use the caboose board tag as SP/ROT artifact names (#3898)
As a follow-up to #3887, use the artifact's `name` field for the board
field for SP/ROT artifacts. This will allow Nexus / sled-agent /
whatever to make decisions on which artifacts to fetch from the
repository in a world where the control plane is accessing a TUF repo
over HTTPS.

The original intent of the `name` field was to provide some key for the
control plane to make decisions on. In the original update system mockup
this was for zone names (i.e. your `kind` would be
`KnownArtifactKind::Zone`, and then your `name` would be something like
`omicron-nexus`). Given hindsight I probably would have named this
something like `key`, or `identifier`, or whatever. (Maybe we still
ought to?)

I had completely forgotten that this is what `name` was for, so I've
written a comment for future me, and hopefully others, explaining it in
`common/src/update.rs`. (For a while I was convinced I needed to plumb
an optional `board` field all the way through the control plane!)

This doesn't matter to Wicket today, as it has the entire repository
present and can read cabooses directly, and that is probably a safer
fallback option anyway.