[wicket] Add `wicket preflight uplink` check (#3929)
The goal of this is to aid in initial installation: we expect to see
various issues with uplink connections and configuration (at multiple
layers: can we get an L1 optical link? once we have an L1 link, can we
talk to the external DNS/NTP servers?) that can impact rack setup in a
way that is difficult to diagnose and debug while rack setup is running.
Instead, _prior to running rack setup_, we can run this new `ssh -t
wicket@... preflight uplink` check: it will walk through setting up the
uplink and talking to the DNS and NTP servers more or less the same way
a person (or RSS) would.

There are (at least) two major shortcomings with the implementation of
this PR:

1. This is entirely an external / CLI-driven check; ideally we could
drive this from within the wicket TUI.
2. It only performs an uplink check for the _current switch_. To check
both switches, an operator must upload the RSS config to both wickets
and then ssh into each wicket to run this check. This isn't hard but is
annoying; ideally ssh'ing into one wicket would run checks on both
scrimlets.

But I think this is already useful as-is.

This covers a portion (but not the entirety of) #3198.