Fix ensuring snapshot creation and deletion (#3947)
If `project_ensure_snapshot` is passed a snapshot where the project and
name matches a record that exists already, but the ID is different, then

    .on_conflict((dsl::project_id, dsl::name))
    .filter_target(dsl::time_deleted.is_null())
    .do_update()
    .set(dsl::time_modified.eq(dsl::time_modified)),

will incorrectly modify the existing record, leading callers of that
function to erroneously believe they were the one to create that
snapshot record. Change `project_ensure_snapshot` to check if an
existing snapshot record has a matching project and name, and return
ObjectAlreadyExists if so.

It's not appropriate to delete a snapshot in any state: if it's in state
Creating, then a `snapshot_create` saga is working on it and
concurrently running `snapshot_delete` will cause a conflict. Add an
argument to `project_delete_snapshot` that describes the states that a
snapshot can be deleted in, depending on on the call site.